requirements.txt
# Core Django
Django
Pillow
django-widget-tweaks
python-dotenv

# Production server
gunicorn
whitenoise

# Database
psycopg2-binary
dj-database-url

# Redis and caching
redis
django-redis

# Background tasks
celery

# API and WebSockets
djangorestframework
django-cors-headers
channels
channels-redis
daphne

# DICOM processing
pydicom
pynetdicom
SimpleITK
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
gdcm
highdicom

# Image processing
opencv-python
scikit-image
matplotlib

# AI and machine learning
numpy
scipy
pandas
torch
torchvision
scikit-learn
transformers

# Document processing
PyMuPDF
python-docx
openpyxl

# Security and authentication
cryptography
PyJWT

# Utilities
python-magic
requests
urllib3
qrcode

# Extensions
django-extensions

# Production monitoring (for new features)
psutil
schedule
gevent

# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
pyjwt==2.8.0

# Printing support for DICOM images
reportlab
# pycups - commented out due to installation issues on some systems
# python-escpos - commented out due to dependency on pycups